change cost from double to int

+++ db/migrate/20150413153157_add_quanity_otf_cost_to_line_item.rb
@@ -2,7 +2,7 @@ class AddQuanityOtfCostToLineItem < ActiveRecord::Migration
   def change
     add_column :line_items, :quantity, :integer
     add_column :line_items, :otf, :boolean
-    add_column :line_items, :cost, :double
+    add_column :line_items, :cost, :integer
     add_column :line_items, :protocol_id, :integer
   end
 end
